% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_functions.R
\name{aggregate_data_and_draws}
\alias{aggregate_data_and_draws}
\title{Aggregate data and draws}
\usage{
aggregate_data_and_draws(
  in_data,
  num_field,
  denom_field,
  draw_fields,
  group_fields = NULL,
  summarize = TRUE
)
}
\arguments{
\item{in_data}{Input data.table, including draws (in rate space), the
numerator and denominator (in count space), and any grouping fields}

\item{num_field}{Field containing the data numerator (e.g. deaths)}

\item{denom_field}{Field containing the data denominator (e.g. population)}

\item{draw_fields}{Character vector of fields containing predictive draws, in
rate space (e.g. mortality rates)}

\item{group_fields}{[optional, default NULL] Character vector containing all
fields for grouping observations during the aggregation. If this field is
empty, all data will be aggregated to a single row and returned with no
identifiers}

\item{summarize}{[bool, default TRUE] should summary columns be added for the
aggregated draws?}
}
\description{
Aggregate the draws (rates) and the data (population and deaths)
  by a set of grouping identifiers
}
