% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_functions.R
\name{aggregate_summarize_excess}
\alias{aggregate_summarize_excess}
\title{Aggregate and summarize excess mortality estimates}
\usage{
aggregate_summarize_excess(
  baseline_deaths_dt,
  aggregate = FALSE,
  group_cols = NULL,
  baseline_draw_cols = paste0("draw", 1:1000),
  obs_death_col = "deaths",
  alpha = 0.05
)
}
\arguments{
\item{baseline_deaths_dt}{data.table containing draws of baseline death
COUNTS (not rates), observed death counts, and optional identifier columns
for the period when excess mortality is being calculated}

\item{aggregate}{[bool, default FALSE] should the data be aggregated across
grouping columns?}

\item{group_cols}{[char, default NULL] If `aggregate` is TRUE, group by these
columns when aggregating baseline and observed deaths. If aggregation is
specified and this argument is left as NULL, the aggregation will occur
across all rows}

\item{baseline_draw_cols}{[character, default c('draw1', ... 'draw1000')]
columns in `baseline_deaths_dt` that contain the draws for predicted
baseline death counts}

\item{obs_death_col}{[character, default 'deaths'] name of the column in
`baseline_deaths_dt` containing observed deaths by week}

\item{alpha}{[numeric, default 0.05] Cutoff for two-tailed uncertainty
interval. For example, the default of 0.05 will use the 95% uncertainty
interval for each outcome}
}
\value{
data.table containing at least the following fields:
   - Any grouping columns, if specified
   - `obs_death_col` as specified in the function argument
   - 'bl_<mean/lower/upper>': Mean and UI bounds for baseline death counts
   - 'smr_<mean/lower/upper>': Mean and UI bounds for Standardized Mortality
       Ratios, calculated as observed / baseline
   - 'ex_d_<mean/lower/upper>': Mean and UI bounds for Excess Deaths,
       calculated as observed - baseline
   - 'sig_under1': Did the observed deaths fall BELOW the baseline deaths UI?
   - 'sig_over1': Did the observed deaths fall ABOVE the baseline deaths UI?
}
\description{
Given draws of baseline death COUNTS (not rates) and observed
  death counts, calculate summary metrics of excess mortality, optionally
  grouping by fields in the data
}
