% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_functions.R
\name{array_split}
\alias{array_split}
\title{Split a matrix into N chunks by row}
\usage{
array_split(data, n_chunks)
}
\arguments{
\item{data}{Matrix or data.frame to split by row}

\item{n_chunks}{[integer] Number of chunks to split the table into}
}
\value{
List of length `n_chunks`, each containing a sub-matrix or data.frame
}
\description{
Helper function to split a matrix into N evenly-sized chunks.
}
\details{
Pulled from StackOverflow: https://bit.ly/3rksN2N
}
