% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{assign_seasonality_ids}
\alias{assign_seasonality_ids}
\title{Assign seasonality grouping IDs}
\usage{
assign_seasonality_ids(input_data, grouping_fields = NULL)
}
\arguments{
\item{input_data}{Data.table of the full input dataset. Must contain fields
for any grouping categories that are assigned}

\item{grouping_fields}{[optional, default NULL] character vector listing any
fields in the dataset that should be used to create unique seasonality
groups}
}
\value{
integer vector with the same number of rows as `input_data` assigning
  holdout IDs.
}
\description{
Create a vector that assigns a (zero-indexed) seasonality
  grouping ID based on other categories in the data. If no categories are
  assigned, all rows get the same seasonality grouping ID (0). If categories
  are based on other ID fields, then a group is assigned to every unique
  combination of values in those fields. For example, if grouings were
  assigned based on the "age" and "location" fields, and the data contained
  four age groupings and 10 unique locations, then 20 unique seasonality
  groupings would be assigned.
}
