% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_data_prep_functions.R
\name{backfill_input_data}
\alias{backfill_input_data}
\title{Backfill missing historical data}
\usage{
backfill_input_data(input_data, index_field, check_vals)
}
\arguments{
\item{input_data}{Input data.table containing input data. The field `year`
(int) is expected}

\item{index_field}{Field containing unique geographic identifiers by location}

\item{check_vals}{Values of the index field to check for backfilling}
}
\value{
Updated data.table with values for Sud Sardegna across all years
}
\description{
Fill missing historical estimates of data using the earliest
  available data value.
}
\details{
This fix is relevant for administrative boundaries that have changed
  over time. At the province level, this is relevant for Sud Sardegna
  province, which is generally listed as its previous component provinces
  through 2016. Sud Sardegna is a composite of territories from several
  now-defunct provinces, including the full territories of Medio Campidano
  and Carbonia-Iglesias provinces. Because there is not an exact
  correspondence with provinces before 2017, replicate the oldest available
  data for Sud Sardegna to previous years.

  The same issue applies to data preparation for the communes (comuna), which
  have less stable boundaries than regions and provinces. Currently in the
  data preparation workflow, preparing population is required to match the
  subset of communes available for deaths.
}
