% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_validity_functions.R
\name{calculate_rmse_rse}
\alias{calculate_rmse_rse}
\title{Calculate RMSE and RSE for predictive estimates}
\usage{
calculate_rmse_rse(
  in_data,
  num_field,
  denom_field,
  est_field,
  group_fields = NULL
)
}
\arguments{
\item{in_data}{Input data.table}

\item{num_field}{Numerator field for the observed data}

\item{denom_field}{Denominator field for the observed data}

\item{est_field}{Estimator field, presented as a rate (num/denom)}

\item{group_fields}{[optional, default NULL] If the predictive validity
metrics should be grouped, list the fields to group them by here. If NULL
(the default), the predictive validity metrics will be calculated across
the entire dataset}
}
\value{
A data.table with the following fields:
  - 'rmse': Root mean squared error, unweighted
  - 'rmse_weighted': Root mean squared error, weighted by population size
  - 'rse': Relative squared error, unweighted
  - 'rse_weighted': Relative squared error, weighted by population size
  - Any grouping columns specified in the function arguments
}
\description{
Given a dataset, calculate Root Mean Squared Error and Relative
 Squared Error (both weighted and unweighted) for predictions across a
 dataset. The observed data should come in the form of separate numerator and
 denominator columns, while the estimates should be a single field estimating
 a rate (i.e. already normalized by denominator)
}
