% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{check_covar_validity}
\alias{check_covar_validity}
\title{Validate prepared covariate data}
\usage{
check_covar_validity(
  prepped_covar,
  covar_name,
  covar_indices,
  model_years,
  location_table
)
}
\arguments{
\item{prepped_covar}{data.table containing the prepped covariate}

\item{covar_name}{Short name for the covariate. This will be used to send the
covariate to a particular prep function.}

\item{covar_indices}{Vector of identifiers that should be used to merge onto
the modeling dataset. The prepared covariate dataset should only include
identifier columns and the covariate value, specified by the covariate name}

\item{model_years}{Vector of years to consider for modeling}

\item{location_table}{Location code merge table for Italy}
}
\value{
NULL (fails if there are any issues in prepared covariate data)
}
\description{
Given prepared covariate data, validate that the correct fields
  are present and that no data is missing
}
