% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_functions.R
\name{create_age_groups}
\alias{create_age_groups}
\title{Create age group categories}
\usage{
create_age_groups(age_cutoffs = seq(0, 100, by = 20))
}
\arguments{
\item{age_cutoffs}{Vector of lower bounds for each age group, starting at 0}
}
\value{
Data.table containing age group categories
}
\description{
Given a set of age cutoffs starting at 0, this function deterministically
creates a data.table of age groupings with IDs and lower/upper bounds for
each age group
}
