% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_data_prep_functions.R
\name{create_template_dt}
\alias{create_template_dt}
\title{Create template data.table to store all observed data}
\usage{
create_template_dt(model_years, model_week_range, location_table, age_groups)
}
\arguments{
\item{model_years}{Vector of years to be modeled}

\item{model_week_range}{Vector of weeks IDs (1-52) to include in the model}

\item{location_table}{Table with `location_code`s for all locations to be
modeled}

\item{age_groups}{Age group index, created by `create_age_groups()`}
}
\value{
Template data.table containing all combinations of these indexes
}
\description{
Create a data table containing all location-age-sex-year-week
  groups that will be estimated in the model. This template can then be
  left-joined to death, population, and covariate data.
}
