% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_functions.R
\name{excess_draws}
\alias{excess_draws}
\title{Get draws of excess deaths}
\usage{
excess_draws(death_data, baseline_draws, template_dt)
}
\arguments{
\item{death_data}{Dataset containing the true number of deaths by location,
year, week, and age group. The true deaths will be matched to the predicted
counterfactual deaths (by way of the `template_dt`) and one subtracted from
the other to estimate excess.}

\item{baseline_draws}{Predicted draws for the number of deaths that would
occur in the absence of COVID}

\item{template_dt}{The template data.table linking predictive draws}
}
\value{
Named list containing two items:
  - 'obs_deaths': Data.table of observed deaths
  - 'excess_draws': Matrix containing excess mortality draws by 
      location/year/week/age. Each row of this matrix corresponds to a row in
      `obs_deaths`.
}
\description{
The traditional calculation for COVID excess mortality 
  calculates excess deaths by subtracting the observed number of deaths from
  an estimated baseline. This function subtracts the observed value (a
  single value) from the baseline (by draw) to get draws of excess.
}
