% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_prep_functions.R
\name{extend_covar_time_series}
\alias{extend_covar_time_series}
\title{Extend a covariate time series forward}
\usage{
extend_covar_time_series(covar_data, model_years)
}
\arguments{
\item{covar_data}{data.table containing covariates with a 'year' field}

\item{model_year}{Vector of years to be included in modeling}
}
\value{
covariate data.table, with years extended to the most recent year of
  data
}
\description{
Helper function to extend a covariate time series to the last
  year modeled by copying forward the most recent year of data available.
}
