% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_functions.R
\name{generate_stwa_draws}
\alias{generate_stwa_draws}
\title{Create post-estimation predictive draws}
\usage{
generate_stwa_draws(
  tmb_sdreport,
  num_draws,
  covariate_names,
  template_dt,
  fourier_harmonics_level
)
}
\arguments{
\item{tmb_sdreport}{output of `TMB::sdreport()` on the fitted model object.
Should include a joint precision matrix (by specifying
`getJointPrecision = TRUE` in the call to `sdreport()`). This object will be
parsed to check for fixed effects, random effects, and the Fourier time
series terms.}

\item{num_draws}{[int] How many posterior predictive samples to take?}

\item{covariate_names}{[char] All covariate field names, including 'intercept'}

\item{template_dt}{[data.table] table containing at least the following fields:
- idx_age: Zero-indexed age grouping
- idx_week: Zero-indexed week grouping (typically week - 1)
- idx_year: Zero-indexed year grouping (typically year - min(year))
- idx_loc: Zero-indexed location grouping
- idx_fourier: Indexer for fourier analysis}

\item{fourier_harmonics_level}{[int, default NULL] Number of levels used to
fit seasonality in the model. This parameter will be ignored if there are
no `Z_fourier` parameters in the fitted output}
}
\value{
A named list with two items:
  - XX TODO
}
\description{
Given the outputs from a fitted TMB model object, create
  an object with posterior predictive draws for all groupings specified by a
  template data.table
}
