% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{get_covar_prep_function}
\alias{get_covar_prep_function}
\title{Pull a specific prep function for an Italy covariate}
\usage{
get_covar_prep_function(covar_name)
}
\arguments{
\item{covar_name}{Short name for the covariate.}
}
\value{
Name of the function to prepare this covariate
}
\description{
Helper function for \code{\link{prepare_covariate}} that checks
  whether a covariate-specific prep function exists yet by searching the
  package namespace. All covariate-specific prep functions should take the
  name format "ita_prep_covar_<COVAR NAME>". Errors if the function does not
  exist; returns the function name if the function does exist.
}
