% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_diagnostic_functions.R
\name{get_covar_vif}
\alias{get_covar_vif}
\title{Get Variance Inflation Factors for a set of covariates}
\usage{
get_covar_vif(in_data, covar_names)
}
\arguments{
\item{in_data}{Input dataset, in data.table format}

\item{covar_names}{[char] vector of covariate names, each of which should
appear as columns in the dataset}
}
\value{
data.table with two fields: 'covar_name' (representing the covariate
  name) and 'vif' (giving the variance inflation factor for each covariate).
}
\description{
Get Variance Inflation Factors (VIFs), indicating
  multicollinearity associated with each covariate in a dataset. A VIF value
  over ~5 indicates high collinearity and might warrant exclusion from a
  dataset
}
