% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_functions.R
\name{get_fourier_seasonality_fit}
\alias{get_fourier_seasonality_fit}
\title{Generate weekly fluctuations in mortality based on fit harmonics}
\usage{
get_fourier_seasonality_fit(fourier_coefs)
}
\arguments{
\item{fourier_coefs}{A matrix with 2 * N rows, where N is the level of
harmonics used to fit seasonality. The first of each pair of terms is the
coefficient on `sin(f(x))`, and the second is the coeffient on `cos(f(x))`,
where `f(x) = (level) * (yearly harmonic frequency) * (week)`. Each column represents
one posterior draw of the same parameter}
}
\value{
Matrix with 52 rows (corrected number of weeks in a modeled year) and the same
  number of columns as `fourier_coefs`, with seasonality fit for each week and draw
}
\description{
This function takes fit coefficients for `sin(.)` and `cos(.)`
  from fourier analysis in TMB and returns the fitted estimates of
  seasonality for each week.
}
