% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_functions.R
\name{index_populated_grid_cells}
\alias{index_populated_grid_cells}
\title{Get populated grid cells by location}
\usage{
index_populated_grid_cells(pop, polys_sf, loc_field)
}
\arguments{
\item{pop}{Population rasterLayer (representing a single year)}

\item{polys_sf}{sf object representing polygons, with a field that has a
unique identifier for each location}

\item{loc_field}{Location identifier field in the polygons sf object}
}
\value{
A data.table containing four columns, where each row represents a
   grid cell:
    - <loc_field>: The location identifier
    - pop: Population
    - lat: Latitude
    - lon: Longitude
}
\description{
Pull the latitude, longitude, and population for each grid cell
  associated with each location in a shapefile
}
