% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_functions.R
\name{ita_prep_sud_sardegna_fix}
\alias{ita_prep_sud_sardegna_fix}
\title{Sud Sardegna fix for Italy input data}
\usage{
ita_prep_sud_sardegna_fix(covar_data)
}
\arguments{
\item{covar_data}{Input data.table containing input data. The fields `icode`
(char) and `year` (int) are expected}
}
\value{
Updated data.table with values for Sud Sardegna across all years
}
\description{
Apply a fix to fill estimates for Sud Sardegna province, which
  is generally listed as its previous component provinces through 2016. Sud
  Sardegna is a composite of territories from several now-defunct provinces,
  including the full territories of Medio Campidano and Carbonia-Iglesias
  provinces. Because there is not an exact correspondence with provinces
  before 2017, replicate the oldest available data for Sud Sardegna to
  previous years. This fix should be noted in the README and project
  appendix.
}
