% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{ita_prepare_covar_elevation}
\alias{ita_prepare_covar_elevation}
\title{Prepare covariate: Elevation}
\usage{
ita_prepare_covar_elevation(covar_fp, pop_raster, polys_sf, projection)
}
\arguments{
\item{covar_fp}{Filepath to the raw covariate}

\item{pop_raster}{[optional] Population rasterBrick object, with one layer
per modeling year. Only used to prepare raster covariates, default NULL.}

\item{polys_sf}{[optional] Polygon boundaries}

\item{projection}{[optional] Character vector giving the proj4 CRS
definition. Example: "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs"}
}
\value{
A list of two items:
  - "prepped_covar": data.table containing the prepped covariate
  - "covar_indices": Vector of identifiers that should be used to merge onto
    the modeling dataset. The prepared covariate dataset should only include
    identifier columns and the covariate value, specified by the covariate
    name
}
\description{
Covariate-specific prep function for elevation by location.
  Specifically, this aggregated covariate represents the population-weighted
  mean elevation inhabited by residents of each province.
}
