% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{ita_prepare_covar_pop_density}
\alias{ita_prepare_covar_pop_density}
\title{Prepare covariate: population density}
\usage{
ita_prepare_covar_pop_density(covar_fp, model_years, location_table, polys_sf)
}
\arguments{
\item{covar_fp}{Filepath to the raw covariate}

\item{model_years}{Vector of years to consider for modeling}

\item{location_table}{Location code merge table for Italy}

\item{polys_sf}{[optional] Polygon boundaries}
}
\value{
A list of two items:
  - "prepped_covar": data.table containing the prepped covariate
  - "covar_indices": Vector of identifiers that should be used to merge onto
    the modeling dataset. The prepared covariate dataset should only include
    identifier columns and the covariate value, specified by the covariate
    name
}
\description{
Covariate-specific prep function for all-ages population density
  by Italian province
}
