% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{ita_prepare_covar_tax_brackets}
\alias{ita_prepare_covar_tax_brackets}
\title{Prepare covariate: proportion of households with income under E10k}
\usage{
ita_prepare_covar_tax_brackets(covar_fp, model_years, location_table)
}
\arguments{
\item{covar_fp}{Filepath to the raw covariate}

\item{model_years}{Vector of years to consider for modeling}

\item{location_table}{Location code merge table for Italy}
}
\value{
data.table containing prepared covariate data and vector of indices
}
\description{
Covariate-specific prep function for proportion of households
  with taxable income under 10,000 Euros per year. A related presentation of
  this data is mean taxable income per household, prepared in another
  function. This function is meant specifically for use with IStat data.
}
