% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{ita_prepare_covar_tax_income}
\alias{ita_prepare_covar_tax_income}
\title{Prepare covariate: mean taxable income across all households}
\usage{
ita_prepare_covar_tax_income(covar_fp, model_years, location_table)
}
\arguments{
\item{covar_fp}{Filepath to the raw covariate}

\item{model_years}{Vector of years to consider for modeling}

\item{location_table}{Location code merge table for Italy}
}
\value{
data.table containing prepared covariate data and vector of indices
}
\description{
Covariate-specific prep function for average taxable income
  across all households in a province. This covariate is derived from two
  IStat datasets, one listing the number of households per tax bracket and
  another listing total taxed income by bracket. A related presentation of
  this data is proportion of households with less than 10k Euros in taxable
  income, prepared in the function `ita_prepare_covar_tax_brackets()`.
}
