% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{ita_prepare_covar_unemp}
\alias{ita_prepare_covar_unemp}
\title{Prepare unemployment covariate}
\usage{
ita_prepare_covar_unemp(covar_data, model_years, location_table)
}
\arguments{
\item{covar_data}{data.table containing the raw covariate data}

\item{model_years}{Vector of years to consider for modeling}

\item{location_table}{Location code merge table for Italy}
}
\value{
data.table containing prepared covariate data and vector of indices
}
\description{
Covariate-specific prep function for unemployment by sex and by
  quarter. This is a convenience function for specific use with datasets
  exported from IStat.
}
