% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{ita_prepare_covar_year_cov}
\alias{ita_prepare_covar_year_cov}
\title{Prepare year covariate}
\usage{
ita_prepare_covar_year_cov(model_years)
}
\arguments{
\item{model_years}{Vector of years to consider for modeling}
}
\value{
data.table containing prepared covariate data and vector of indices
}
\description{
Covariate-specific prep function for the year covariate. This
  one is pretty simple - the (unnormalized) covariate is the same as the year
}
