% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{ita_prepare_covariate}
\alias{ita_prepare_covariate}
\title{Prepare an Italy covariate for modeling}
\usage{
ita_prepare_covariate(
  covar_name,
  covar_fp,
  model_years,
  location_table,
  pop_raster = NULL,
  polys_sf = NULL,
  projection = NULL,
  final_obs_week = NULL
)
}
\arguments{
\item{covar_name}{Short name for the covariate. This will be used to send the
covariate to a particular prep function.}

\item{covar_fp}{Filepath to the raw covariate}

\item{model_years}{Vector of years to consider for modeling}

\item{location_table}{Location code merge table for Italy}

\item{pop_raster}{[optional] Population rasterBrick object, with one layer
per modeling year. Only used to prepare raster covariates, default NULL.}

\item{polys_sf}{[optional] Polygon boundaries}

\item{projection}{[optional] Character vector giving the proj4 CRS
definition. Example: "+proj=utm +zone=32 +datum=WGS84 +units=m +no_defs"}

\item{final_obs_week}{[optional] ID for the final observed week in 2020,
formatted as an integer between 1 and 52 inclusive.}
}
\value{
A list of two items:
  - "prepped_covar": data.table containing the prepped covariate
  - "covar_indices": Vector of identifiers that should be used to merge onto
    the modeling dataset. The prepared covariate dataset should only include
    identifier columns and the covariate value, specified by the covariate
    name
}
\description{
Generic function to prepare a covariate for modeling. This
  function is a wrapper for sub-functions called for specific covariates and
  will fail if a non-initialized covariate has been passed. It returns both
  the prepared covariate and a vector of expected indices that can be used
  to merge back onto the original dataset.
}
