% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_functions.R
\name{ita_prepare_deaths}
\alias{ita_prepare_deaths}
\title{Prepare death data for Italy}
\usage{
ita_prepare_deaths(
  deaths_raw,
  age_cutoffs,
  model_years,
  first_covid_death_date
)
}
\arguments{
\item{deaths_raw}{Data.table downloaded from IStat}

\item{age_cutoffs}{Vector of the starting years for each age group bin. For
more information, see \link{\code{create_age_groups}}}

\item{model_years}{Vector of years to be included in the mortality model, up
to and including the current year}

\item{first_covid_death_date}{R Date object designating the date of the first
confirmed COVID death in Italy.}
}
\value{
Data.table of formatted death data
}
\description{
Prepare death data for Italy
}
