% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_data_prep_functions.R
\name{ita_prepare_deaths_single_year}
\alias{ita_prepare_deaths_single_year}
\title{Prepare a single year of death data for Italy}
\usage{
ita_prepare_deaths_single_year(
  deaths_raw,
  age_cutoffs,
  this_year,
  first_covid_death_date
)
}
\arguments{
\item{deaths_raw}{Data.table downloaded from IStat}

\item{age_cutoffs}{Vector of the starting years for each age group bin. For
more information, see \link{\code{create_age_groups}}}

\item{this_year}{Year to be extracted}

\item{first_covid_death_date}{R Date object designating the date of the first
confirmed COVID death in Italy.}
}
\value{
Data.table of formatted death data
}
\description{
Prepare a single year of death data for Italy. This function is
  wrapped by `ita_prepare_deaths_all_years()`.
}
