% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_data_prep_functions.R
\name{ita_prepare_pop}
\alias{ita_prepare_pop}
\title{Prepare population data for Italy}
\usage{
ita_prepare_pop(pop_raw, age_cutoffs)
}
\arguments{
\item{pop_raw}{Raw population data.table downloaded from IStat}

\item{age_cutoffs}{Vector of the starting years for each age group bin. For}
}
\value{
Data.table of formatted population data
}
\description{
Given raw population data by age, sex, year, and province in
  Italy, format the data for modeling. This is a convenience function
  specifically designed to be use with IStat data. For more information, see
  the repository README.
}
