% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_data_prep_functions.R
\name{ita_read_istat_csv}
\alias{ita_read_istat_csv}
\title{Read CSV downloaded from iStat}
\usage{
ita_read_istat_csv(fp)
}
\arguments{
\item{fp}{[char] The filepath of the input file to read}
}
\value{
A data.table representation of the CSV, with NAs appropriately placed
}
\description{
Read a CSV downloaded from iStat, the Italian Statistical
  Agency, as an R data.table
}
