% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_covariate_prep_functions.R
\name{ita_temperature_query_api}
\alias{ita_temperature_query_api}
\title{Load and cache daily temperature data from API}
\usage{
ita_temperature_query_api(
  model_years,
  api_key,
  prov_latlong,
  cache_file,
  refresh_cache = FALSE
)
}
\arguments{
\item{model_years}{Years of data to pull}

\item{api_key}{API key for querying the Meteostat API}

\item{prov_latlong}{Table indexing locations to latitute and longitude. Must
contain at least the following three fields:
 - 'location_code': Numeric code identifying location
 - 'lat': Latitude
 - 'lon': Longitude}

\item{cache_file}{File where the pulled data will be cached}

\item{refresh_cache}{[optional, default FALSE] Should the data be re-pulled
even if a cache file exists?}
}
\value{
Data.table containing full Meteostat daily API data as well as
 location code for all available years
}
\description{
Pull daily temperature data from the Meteostat API. This
  function uses cacheing: if the `cache_file` already exists, it pulls the
  results from file rather than re-querying the API
}
