% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{italian_cities_dt}
\alias{italian_cities_dt}
\title{Get largest Italian cities}
\usage{
italian_cities_dt()
}
\value{
Data.table with three fields: 'name', 'lon', 'lat'
}
\description{
Returns a data.table with all Italian cities with an estimated
  population of over 500,000
}
