% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_functions.R
\name{load_format_pop_raster}
\alias{load_format_pop_raster}
\title{Load and format population raster}
\usage{
load_format_pop_raster(pop_fp_format, model_years, country_polys, projection)
}
\arguments{
\item{pop_fp_format}{Each of the worldpop rasters are downloaded separately
by year. This format specifies the location of each raster, with the
'{year}' standing in for a particular year. For example:
`pop_filepath_format = '/path/to/pop/raster/my_raster_{year}.tif'`}

\item{model_years}{The years needed for the model. The function will load
separate raster layers for each model year based on the filepath template}

\item{country_polys}{The spatial polygons object (in sf format) that the
raster will be clipped to}

\item{projection}{Projection used for spatial data throughout the model}
}
\value{
rasterBrick clipped to the polygons object
}
\description{
Load a population raster from file, clip it to the given
  shapefile, and return as a raster brick for the years specified
}
