% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_functions.R
\name{load_format_spatial_polys}
\alias{load_format_spatial_polys}
\title{Load and format spatial polygons for modeling}
\usage{
load_format_spatial_polys(shp_in_fp, abbrev_field, location_table)
}
\arguments{
\item{shp_in_fp}{Input filepath to a spatial object (usually a shapefile)}

\item{abbrev_field}{Field listing two-character province abbreviations}

\item{location_table}{Data.frame listing province information. Should contain
at least the following fields:
  - 'abbrev': Two-character province abbreviations (eg 'TO', 'VC')
  - 'location_code': Province codes that will be assigned to the spatial
       polygons object}
}
\value{
List with two items:
  - 'shp_sf': Formatted spatial polygons, in SF format
  - 'shp_sp': Formatted spatial polygons, in SP format
}
\description{
Load a spatial polygons object in SF format, assign location
  codes based on each province abbreviation, and return polygons objects in
  both SF and SP formats
}
