% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{map_ita_choropleth_region}
\alias{map_ita_choropleth_region}
\title{Make a choropleth map of Italian REGIONS}
\usage{
map_ita_choropleth_region(
  region_sf,
  in_data,
  map_field,
  fill_list,
  fill_lims = NULL,
  titles_list = list(),
  fill_type = "continuous",
  show_legend = TRUE,
  labels_dt = NULL,
  save_fp = NULL,
  file_height = 7.25,
  file_width = 6
)
}
\arguments{
\item{region_sf}{[sf] Spatial object of regions. Will be added as an outline
to map}

\item{in_data}{[data.table] Dataset to be plotted. Should only include one
row of data per region}

\item{map_field}{[char] Field in `in_data` to be mapped}

\item{fill_list}{list of fill parameters that should be used, in the format
of arguments to the `scale_fill_<fill_type>`}

\item{fill_lims}{[numeric, default NULL] Vector of upper and lower values for
color scale.}

\item{titles_list}{[default empty list] list of any titles that should be
included, in the format of arguments to the ggplot `labs` command}

\item{fill_type}{[optional, default 'continuous'] Scale fill type eg.
'discrete' or 'continuous'}

\item{show_legend}{Should the legend be featured on the plot?}

\item{labels_dt}{[optional] Data.table used to label locations on the graph. Should be
a data.table containing at least the fields `label_name`, `x`, `y`, `x1_lseg`,
`x2_lseg`, `y1_lseg`, `y2_lseg`}

\item{save_fp}{[optional, default NULL] if not NULL (the default), saves to
file rather than returning an object. If filled, should be a PNG extension}

\item{file_height}{[optional, default 8] map height in inches}

\item{file_width}{[optional, default 6] map width in inches}
}
\value{
Either returns a ggplot object or, if `save_fp` is not NULL, saves to
  file and returns NULL
}
\description{
Create a choropleth map for Italian regions based on
  continuous input data by region
}
