% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{normalize_adfun}
\alias{normalize_adfun}
\title{Run TMB normalization}
\usage{
normalize_adfun(adfun, flag, verbose = FALSE)
}
\arguments{
\item{adfun}{The ADFunction to normalize}

\item{flag}{[char] flag to indicate when random effects only have been
incorporated into the joint negative log-likelihood}

\item{verbose}{[bool, default FALSE] return a message about normalization?}
}
\value{
Normalized ADFunction
}
\description{
Get the TMB normalization constant from random effects
}
