% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_prep_functions.R
\name{rescale_prepped_covariates}
\alias{rescale_prepped_covariates}
\title{Rescale observed covariates}
\usage{
rescale_prepped_covariates(
  input_data,
  covar_names,
  subset_field = NULL,
  subset_field_values = NULL
)
}
\arguments{
\item{input_data}{Data.table of the full dataset containing observations
along with covariate data}

\item{covar_names}{[char] vector of covariate names. All of these should be
included as fields in the `input_data`.}

\item{subset_field}{[char, optional] Fill this option to rescale based on
only observations from the training data (not the test data). This
argument gives the field to subset on}

\item{subset_field_values}{[optional] Fill this option to rescale based on
only observations from the training data (not the test data). This argument
gives the values of `subset_field` indicating that an observation was in
the training dataset and should be used for rescaling}
}
\value{
Named list containing two data.tables:
    - 'data_rescaled': Prepared dataset with rescaled covariate values
    - 'covariate_scaling_factors': Mean and SD of the original dataset (use
         to back-transform the scaled covariates)
}
\description{
Rescale covariates so that each covariate has mean zero and
  variance 1 when measured at all observations in the dataset. Optionally,
  this function allows for rescaling based only on observations in the
  training subset of the data, as denoted by a field in the dataset
}
