% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{Take multivariate normal draws given a mean vector and precision matrix}
\usage{
rmvnorm_prec(mu, prec, n.sims, find_near_pd = FALSE)
}
\arguments{
\item{mu}{vector of parameter means}

\item{prec}{joint precision matrix}

\item{n.sims}{number of draws}
}
\value{
length(mu) by n.sims matrix of parameter draws
}
\description{
Take multivariate normal draws given a mean vector and precision matrix
}
