% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{setup_run_tmb}
\alias{setup_run_tmb}
\title{Set up and run TMB}
\usage{
setup_run_tmb(
  tmb_data_stack,
  params_list,
  tmb_random,
  tmb_map,
  DLL,
  tmb_outer_maxsteps,
  tmb_inner_maxsteps,
  run_symbolic_analysis = FALSE,
  normalize = FALSE,
  set_limits = FALSE,
  limit_max = 10,
  limit_min = -limit_max,
  model_name = "model",
  verbose = FALSE
)
}
\arguments{
\item{tmb_data_stack}{List containing all data inputs expected in the TMB
CPP file}

\item{params_list}{List containing all parameters expected in the TMB CPP file}

\item{tmb_random}{Character vector containing all random effects that will be
optimized in the inner optimizer}

\item{tmb_map}{Named list containing parameters that will be treated in a
particular way by the optimizer}

\item{tmb_outer_maxsteps}{Max number of steps taken by the outer optimizer}

\item{tmb_inner_maxsteps}{Max number of steps taken by the inner optimizer
in a single outer optimizer step}

\item{run_symbolic_analysis}{[boolean, default FALSE] run symbolic analysis
to speed up model run time?}

\item{normalize}{[boolean, default FALSE] Run normalize fix for models with
large random effects sets? Only run this option if you know what you are
doing and have explicitly coded this fix into the TMB code.}

\item{set_limits}{[boolean, default FALSE] Set limits for fixed effects?}

\item{limit_max}{[numeric, default 10] If limits are set in the model,
maximum value for any fixed effect}

\item{limit_min}{[numeric, default -limit_max] If limits are set in the model,
minimum value of any fixed effect}

\item{model_name}{[char, default "model"] name of the model}

\item{verbose}{[boolean, default FALSE] Should this function return}
}
\value{
list of two objects: obj (ADFunction object), and opt (optimized
  nlminb object)
}
\description{
Generic TMB model run handler. Sets up the ADFun object, applies
  model speedups and fixes as specified, and optimizes using `nlminb`. This
  is meant to be a helper function run by more specific model functions.
}
