% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita_data_prep_functions.R
\name{week_id_from_date}
\alias{week_id_from_date}
\title{Get week ID from date}
\usage{
week_id_from_date(date, tryFormats = c("\%Y-\%m-\%d", "\%Y/\%m/\%d"))
}
\arguments{
\item{date}{String representation of a date (singleton or vector). Can also
be an object or vector of type Date, in which case it will not be converted}

\item{tryFormats}{[optional] Formats to try when converting the string
representation of a date into a Date object. Defaults to the default
formats tried by `as.Date()`, "%Y-%m-%d" and "%Y/%m/%d". Not used if the
`date` parameter is already of type Date}
}
\value{
integer vector giving week of the year
}
\description{
Given a text representation of a date, get the 1-indexed ID for
  week of the year associated with that date. Example: '2020-01-01' -> 1
}
