% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation_functions.R
\name{weekly_harmonic_freq}
\alias{weekly_harmonic_freq}
\title{Get weekly harmonic frequency}
\usage{
weekly_harmonic_freq()
}
\value{
A float constant
}
\description{
Get the constant that translates from the unit circle to a
  52-week harmonic cycle. Used for fourier analysis of seasonality.
}
