% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_functions.R
\name{build_adjacency_matrix}
\alias{build_adjacency_matrix}
\title{Build adjacency matrix}
\usage{
build_adjacency_matrix(poly_sp, style = "B", allow_zero_neighbors = TRUE)
}
\arguments{
\item{poly_sp}{A SpatialPolygonsDataFrame object}

\item{style}{One of "B","W","C", or "S". For more information, see the
documentation in the \code{\link{spdep}} package:
https://www.rdocumentation.org/packages/spdep/versions/1.1-3/topics/nb2mat}

\item{allow_zero_neighbors}{[bool, default TRUE] Allow polygons with no
neighbors?}
}
\value{
sparse dsCMatrix representing adjacency between polygons in the
  `poly_sp` object
}
\description{
Builds a matrix describing adjacency between polygons in the
  a spatial object. This function is a wrapper of two functions in the
  \code{\link{spdep}} package.
}
