% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_prep_functions.R
\name{ita_prepare_covariate}
\alias{ita_prepare_covariate}
\title{Prepare an Italy covariate for modeling}
\usage{
ita_prepare_covariate(covar_data, covar_name, model_years, location_table)
}
\arguments{
\item{covar_data}{data.table containing the covariate}

\item{covar_name}{Short name for the covariate. This will be used to send the
covariate to a particular prep function.}

\item{model_years}{Vector of years to consider for modeling}

\item{location_table}{Location code merge table for Italy}
}
\value{
A list of two items:
  - "prepped_covar": data.table containing the prepped covariate
  - "covar_indices": Vector of identifiers that should be used to merge onto
    the modeling dataset. The prepared covariate dataset should only include
    identifier columns and the covariate value, specified by the covariate
    name
}
\description{
Generic function to prepare a covariate pulled from IStat data.
  This function is a wrapper for sub-functions called for specific covariates
  and will fail if a non-initialized covariate has been passed. It returns
  both the prepared covariate and a vector of expected indices that can be
  used to merge back onto the original dataset.
}
