% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{normalize_adfun}
\alias{normalize_adfun}
\title{Run TMB normalization}
\usage{
normalize_adfun(adfun, flag, value = 0, verbose = FALSE)
}
\arguments{
\item{adfun}{The ADFunction to normalize}

\item{flag}{[char] Flagging variable to indicate whether random effects only have been
incorporated into the joint negative log-likelihood}

\item{value}{[int, default 0] Value that the flagging variable takes when the data
should NOT be incorporated into the jnll}

\item{verbose}{[bool, default FALSE] return a message about normalization?}
}
\value{
Normalized ADFunction
}
\description{
Get the TMB normalization constant from random effects
}
