% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{setup_run_tmb}
\alias{setup_run_tmb}
\title{Set up and run TMB}
\usage{
setup_run_tmb(
  tmb_data_stack,
  params_list,
  tmb_random,
  tmb_map,
  tmb_outer_maxsteps,
  tmb_inner_maxsteps,
  normalize = FALSE,
  run_symbolic_analysis = FALSE,
  set_limits = FALSE,
  limit_max = 10,
  limit_min = -limit_max,
  parallel_model = FALSE,
  optimization_method = "nlminb",
  model_name = "model",
  verbose = FALSE,
  inner_verbose = FALSE
)
}
\arguments{
\item{tmb_data_stack}{List containing all data inputs expected in the TMB
CPP file}

\item{params_list}{List containing all parameters expected in the TMB CPP file}

\item{tmb_random}{Character vector containing all random effects that will be
optimized in the inner optimizer}

\item{tmb_map}{Named list containing parameters that will be treated in a
particular way by the optimizer}

\item{tmb_outer_maxsteps}{Max number of steps taken by the outer optimizer}

\item{tmb_inner_maxsteps}{Max number of steps taken by the inner optimizer
in a single outer optimizer step}

\item{normalize}{[boolean, default FALSE] Run TMB's automatic process normalization
function? Adds two additional items to the TMB data stack: `auto_normalize` (1) and
`early_return` (0). Both are used by the \code{\link{normalize_adfun}} function to
get the normalizing constant prior to optimization. For more details, see
\link{\code{TMB::normalize}}.}

\item{run_symbolic_analysis}{[boolean, default FALSE] run symbolic analysis
to speed up model run time?}

\item{set_limits}{[boolean, default FALSE] Set limits for fixed effects?}

\item{limit_max}{[numeric, default 10] If limits are set in the model,
maximum value for any fixed effect}

\item{limit_min}{[numeric, default -limit_max] If limits are set in the model,
minimum value of any fixed effect}

\item{parallel_model}{[bool, default FALSE] Is the model implemented in parallel? If
TRUE, opens multiple OMP threads before fitting}

\item{optimization_method}{[char, default 'nlminb'] Outer optimization method to use
for fitting, implemented in the optimx library. Recommended options include 'nlminb'
and 'L-BFGS-B'}

\item{model_name}{[char, default "model"] name of the model}

\item{verbose}{[boolean, default FALSE] Should this function return logging
information about the stage of model fitting, including the outer optizimer
sampling? This will be somewhat long (>100 lines)}

\item{inner_verbose}{[boolean, default FALSE] Should this function return
logging information about inner optimizer sampling? This can be useful for
debugging, but will show very verbose (>10k lines) output.}
}
\value{
list of two objects: obj (ADFunction object), and opt (optimized
  nlminb object)
}
\description{
Generic TMB model run handler. Sets up the ADFun object, applies
  model speedups and fixes as specified, and optimizes using `nlminb`. This
  is meant to be a helper function run by more specific model functions.
}
